requirements.txt
python-dotenv
python-telegram-bot
pymongo
pymongo[srv]
openpyxl